requirements.txt
rich
gitpython
openai
pydantic
python-dotenv
plotly
requests
pytest
textual
pymongo[srv]
ipykernel